adds runner file to night reader
