+++ lib/night_reader.rb
@@ -1,0 +1,5 @@
+require_relative 'file_io'
+require_relative 'night_read'
+
+night_reader = NightRead.new
+night_reader.encode_file_to_english
