+++ test/night_read_test.rb
@@ -1,0 +1,14 @@
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/night_read'
+
+class NightReadTest < Minitest::Test
+
+  def test_it_exists
+    nr = NightRead.new
+
+    assert_instance_of NightRead, nr
+  end
+
+
+end
